Struct mesh_io::medit::Mesh

pub struct Mesh { /* private fields */ }

MeditMesh data structure.

It stores main mesh informations about geometric coordinates and topology.

Implementations

impl Mesh

pub fn from_file(path: impl AsRef<Path>) -> Result<Mesh, Error>

Import from a medit mesh file.

pub fn from_reader(r: impl BufRead) -> Result<Mesh, Error>

impl Mesh

pub fn write_to<W: Write>(&self, w: W) -> Result<()>

impl Mesh

pub fn new(dimension: usize) -> Self

pub fn from_raw_parts(
dimension: usize,
coordinates: Vec<f64>,
node_refs: Vec<isize>,
topology: Vec<(ElementType, Vec<usize>, Vec<isize>)>
) -> Self

pub fn dimension(&self) -> usize

Returns the dimension of the mesh (2D, 3D, …)

pub fn coordinates(&self) -> &[f64]

Returns a slice of f64 representing the coordinates of the points in the mesh. e.g. [x1, y1, x2, y2, …]

pub fn node_refs(&self) -> &[isize]

Returns the references to the points of the mesh.

Its length is dimension times smaller than the coordinates slice.

pub fn node(&self, idx: usize) -> &[f64]

pub fn nodes(&self) -> impl Iterator<Item = (&[f64], isize)>

pub fn nodes_mut(&mut self) -> impl Iterator<Item = (&mut [f64], &mut isize)>

pub fn topology(&self) -> &[(ElementType, Vec<usize>, Vec<isize>)]

Returns a slice of matrices containing the connectivity for each mesh element

pub fn elements(&self) -> impl Iterator<Item = (ElementType, &[usize], isize)>

pub fn elements_mut(
&mut self
) -> impl Iterator<Item = (ElementType, &mut [usize], &mut isize)>

pub fn node_count(&self) -> usize

Returns the number of nodes (vertices) of the mesh.

pub fn element_count(&self) -> usize

impl Mesh

pub fn duplicate(self, n: usize) -> Mesh

Place copies of self side by side in order to form a grid.

pub fn refine(&self) -> Mesh

Split each element in N smaller elements of the same type, where N is the number of nodes of the element.